refactor: 패키지 및 클래스 명 변경
- LadderController -> Game
- LadderStrategy -> GenerateDirectionStrategy
- RandomLadderStrategy -> RandomStrategy
