+++ README.md
@@ -1,6 +1,6 @@
 # java-ladder
 
-### LadderGame
+### Game
 - [x] 사다리 게임에 사람이 참여한다.
 - [x] 사다리를 생성한다.
 - [x] 사다리 게임 결과물이 사람 수 만큼 있다.
