--- src/main/java/domain/LadderStrategy.java
@@ -1,6 +1,0 @@
-package domain;
-
-interface LadderStrategy {
-
-    boolean creatable();
-}
