+++ src/main/java/ladder/domain/Name.java
@@ -1,4 +1,4 @@
-package domain;
+package ladder.domain;
 
 import java.util.regex.Pattern;
