+++ src/main/java/ladder/domain/ladder/generator/GenerateDirectionStrategy.java
@@ -1,0 +1,6 @@
+package ladder.domain.ladder.generator;
+
+public interface GenerateDirectionStrategy {
+
+    boolean creatable();
+}
