+++ src/main/java/ladder/view/InputView.java
@@ -1,4 +1,4 @@
-package view;
+package ladder.view;
 
 import java.io.BufferedReader;
 import java.io.IOException;
