+++ src/test/java/ladder/domain/NameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package ladder.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
